% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{generateSoccerHeatmap}
\alias{generateSoccerHeatmap}
\title{Generate a Soccer Heatmap}
\usage{
generateSoccerHeatmap(
  data,
  x,
  y,
  tipo = "Tudo",
  jogo = NULL,
  time = NULL,
  jogador = NULL,
  xBins = 18,
  yBins = 18,
  arrow = c("r"),
  colLow = "grey",
  colHigh = "darkred"
)
}
\arguments{
\item{data}{Dados do jogo contendo informações sobre as jogadas.}

\item{x}{Nome da coluna no dataframe que contém as coordenadas x dos eventos.}

\item{y}{Nome da coluna no dataframe que contém as coordenadas y dos eventos.}

\item{tipo}{Tipo de evento para o mapa de calor. Pode ser "Chute", "Passe" ou "Tudo".}

\item{jogo}{ID do jogo a ser filtrado (opcional).}

\item{time}{Nome do time a ser filtrado (opcional).}

\item{jogador}{Nome do jogador a ser filtrado (opcional).}

\item{xBins}{Número de divisões (bins) no eixo x para o mapa de calor.}

\item{yBins}{Número de divisões (bins) no eixo y para o mapa de calor.}

\item{arrow}{Direção das setas a serem usadas no gráfico (opcional).}

\item{colLow}{Cor para os valores baixos de densidade.}

\item{colHigh}{Cor para os valores altos de densidade.}
}
\value{
Um gráfico de mapa de calor gerado pela função \code{soccerHeatmap}.
}
\description{
Esta função gera um mapa de calor de futebol baseado nos eventos de jogo.
}
