% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{generateSoccerMap}
\alias{generateSoccerMap}
\title{Generate a Soccer Map}
\usage{
generateSoccerMap(
  data,
  x,
  y,
  endx = NULL,
  endy = NULL,
  tipo = "Pass",
  jogo = NULL,
  time = NULL,
  jogador = NULL
)
}
\arguments{
\item{data}{Dados do jogo contendo informações sobre as jogadas.}

\item{x}{Nome da coluna no dataframe que contém as coordenadas x dos eventos.}

\item{y}{Nome da coluna no dataframe que contém as coordenadas y dos eventos.}

\item{endx}{Nome da coluna no dataframe que contém as coordenadas x do destino (para passes).}

\item{endy}{Nome da coluna no dataframe que contém as coordenadas y do destino (para passes).}

\item{tipo}{Tipo de evento a ser plotado. Pode ser "Passe", "Chute" ou "Defensivo".}

\item{jogo}{ID do jogo a ser filtrado (opcional).}

\item{time}{Nome do time a ser filtrado (opcional).}

\item{jogador}{Nome do jogador a ser filtrado (opcional).}
}
\value{
Um gráfico de eventos de futebol no campo gerado pela função \code{ggplot2::ggplot}.
}
\description{
Esta função gera um gráfico com os eventos de futebol no campo, incluindo passes e chutes.
}
