% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pitch.R
\name{soccerHeatmap}
\alias{soccerHeatmap}
\title{Plota um mapa de calor no campo de futebol}
\usage{
soccerHeatmap(
  df,
  lengthPitch = 105,
  widthPitch = 68,
  xBins = 10,
  yBins = NULL,
  kde = FALSE,
  arrow = c("none", "r", "l"),
  colLow = "white",
  colHigh = "red",
  title = NULL,
  subtitle = NULL,
  x = "x",
  y = "y"
)
}
\arguments{
\item{df}{DataFrame com as coordenadas dos eventos no campo, com as colunas \code{x} e \code{y} representando as posições.}

\item{lengthPitch}{Comprimento do campo de futebol (padrão 105 metros).}

\item{widthPitch}{Largura do campo de futebol (padrão 68 metros).}

\item{xBins}{Número de bins horizontais para o mapa de calor (padrão 10).}

\item{yBins}{Número de bins verticais para o mapa de calor. Se \code{NULL}, é igual a \code{xBins}.}

\item{kde}{Se \code{TRUE}, plota o mapa de calor utilizando uma estimativa de densidade de Kernel.}

\item{arrow}{Controla a direção das setas no campo, opções: "none", "r", "l" (padrão "none").}

\item{colLow}{Cor do valor mais baixo do mapa de calor (padrão "white").}

\item{colHigh}{Cor do valor mais alto do mapa de calor (padrão "red").}

\item{title}{Título do gráfico.}

\item{subtitle}{Subtítulo do gráfico.}

\item{x}{Nome da coluna que representa a coordenada \code{x} (padrão "x").}

\item{y}{Nome da coluna que representa a coordenada \code{y} (padrão "y").}
}
\value{
Um gráfico do mapa de calor sobre o campo de futebol.
}
\description{
Esta função cria um mapa de calor (zonal ou baseado em densidade de Kernel) sobre o campo de futebol, utilizando coordenadas de localização no campo.
}
